jewellery_shop/requirements.txt
asgiref==3.8.1
Django==4.2.16
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==23.0.0
packaging==24.1
pillow==10.4.0
PyJWT==2.9.0
setuptools==72.1.0
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
uv==0.4.10
wheel==0.44.0
whitenoise==6.7.0
langchain
langchain-google-genai
six
dotenv
